+++ src/Adapter/Pwa/fileFilter.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {(file: string) => boolean} fileFilter
+ */