+++ src/Service/Pwa/Command/GenerateServiceWorkerCommand.mjs
@@ -3,6 +3,8 @@ import { SKIP_WAITING } from "../../../../../flux-pwa-api/src/Adapter/Pwa/SKIP_W
 import { dirname, extname, join, relative } from "node:path/posix";
 import { readdir, readFile, stat, writeFile } from "node:fs/promises";
 
+/** @typedef {import("../../../Adapter/Pwa/fileFilter.mjs").fileFilter} fileFilter */
+
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
 export class GenerateServiceWorkerCommand {
@@ -26,15 +28,21 @@ export class GenerateServiceWorkerCommand {
      * @param {string} application_cache_prefix
      * @param {string | null} service_worker_template_mjs_file
      * @param {{[key: string]: *} | null} data
+     * @param {fileFilter | null} filter_filter
+     * @param {boolean | null} ignore_jsdoc_files
      * @returns {Promise<void>}
      */
-    async generateServiceWorker(web_root, service_worker_mjs_file, application_cache_prefix, service_worker_template_mjs_file = null, data = null) {
+    async generateServiceWorker(web_root, service_worker_mjs_file, application_cache_prefix, service_worker_template_mjs_file = null, data = null, filter_filter = null, ignore_jsdoc_files = null) {
         await writeFile(service_worker_mjs_file, "");
 
         await writeFile(service_worker_mjs_file, (await readFile(service_worker_template_mjs_file ?? join(__dirname, "..", "..", "..", "Adapter", "service-worker-template.mjs"), "utf8")).replaceAll("{ /*%DATA%*/ }", JSON.stringify({
             ...data,
             APPLICATION_CACHE_FILES: [
-                "",
+                ...filter_filter !== null && !filter_filter(
+                    ""
+                ) ? [] : [
+                    ""
+                ],
                 ...await (async function scanFiles(folder) {
                     const files = [];
 
@@ -48,7 +56,15 @@ export class GenerateServiceWorkerCommand {
                                 _file
                             ));
                         } else {
-                            if ([
+                            const web_root_file = relative(web_root, _file);
+
+                            if (filter_filter !== null && !filter_filter(
+                                web_root_file
+                            )) {
+                                continue;
+                            }
+
+                            if ((ignore_jsdoc_files ?? true) && [
                                 ".cjs",
                                 ".js",
                                 ".mjs"
@@ -60,7 +76,7 @@ export class GenerateServiceWorkerCommand {
                                 }
                             }
 
-                            files.push(relative(web_root, _file));
+                            files.push(web_root_file);
                         }
                     }